+++ servicetests/tests/ShadowUpdate/main.cpp
@@ -163,7 +163,7 @@ std::shared_ptr<IotShadowClient> build_mqtt5_client(
     builder->WithConnectOptions(connectOptions);
     if (cmdData.input_port != 0)
     {
-        builder->WithPort(static_cast<uint16_t>(cmdData.input_port));
+        builder->WithPort(static_cast<uint32_t>(cmdData.input_port));
     }
     // Setup lifecycle callbacks
     builder->WithClientConnectionSuccessCallback(
